feat: improves usecase file and response structure
